+++ src/coprocessor/mod.rs
@@ -15,6 +15,7 @@ mod checksum;
 pub mod codec;
 mod dag;
 mod endpoint;
+mod error;
 pub mod local_metrics;
 mod metrics;
 mod readpool_context;
@@ -22,96 +23,16 @@ mod statistics;
 mod util;
 
 pub use self::endpoint::err_resp;
+pub use self::error::{Error, Result};
 pub use self::readpool_context::Context as ReadPoolContext;
 
-use std::error;
-use std::result;
 use std::time::Duration;
 
-use kvproto::{errorpb, kvrpcpb};
-use tipb;
+use kvproto::kvrpcpb;
 
-use storage;
 use util::time::Instant;
 
 const SINGLE_GROUP: &[u8] = b"SingleGroup";
-
-quick_error! {
-    #[derive(Debug)]
-    pub enum Error {
-        Region(err: errorpb::Error) {
-            description("region related failure")
-            display("region {:?}", err)
-        }
-        Locked(l: kvrpcpb::LockInfo) {
-            description("key is locked")
-            display("locked {:?}", l)
-        }
-        Outdated(elapsed: Duration, tag: &'static str) {
-            description("request is outdated")
-        }
-        Full(allow: usize) {
-            description("running queue is full")
-        }
-        Eval(err: tipb::select::Error) {
-            from()
-            description("eval failed")
-            display("eval error {:?}", err)
-        }
-        Other(err: Box<error::Error + Send + Sync>) {
-            from()
-            cause(err.as_ref())
-            description(err.description())
-            display("unknown error {:?}", err)
-        }
-    }
-}
-
-pub type Result<T> = result::Result<T, Error>;
-
-impl From<storage::engine::Error> for Error {
-    fn from(e: storage::engine::Error) -> Error {
-        match e {
-            storage::engine::Error::Request(e) => Error::Region(e),
-            _ => Error::Other(box e),
-        }
-    }
-}
-
-impl From<self::dag::expr::Error> for Error {
-    fn from(e: self::dag::expr::Error) -> Error {
-        Error::Eval(e.into())
-    }
-}
-
-impl From<::util::codec::Error> for Error {
-    fn from(e: ::util::codec::Error) -> Error {
-        let mut err = tipb::select::Error::new();
-        err.set_msg(format!("{}", e));
-        Error::Eval(err)
-    }
-}
-
-impl From<storage::txn::Error> for Error {
-    fn from(e: storage::txn::Error) -> Error {
-        match e {
-            storage::txn::Error::Mvcc(storage::mvcc::Error::KeyIsLocked {
-                primary,
-                ts,
-                key,
-                ttl,
-            }) => {
-                let mut info = kvrpcpb::LockInfo::new();
-                info.set_primary_lock(primary);
-                info.set_lock_version(ts);
-                info.set_key(key);
-                info.set_lock_ttl(ttl);
-                Error::Locked(info)
-            }
-            _ => Error::Other(box e),
-        }
-    }
-}
 
 #[derive(Debug)]
 pub struct ReqContext {
